Create new directory to hold new tests
